♻ split FeedAdd components
