+++ src/components/FeedAdd/ImgAddPreview/styles.js
@@ -1,17 +1,7 @@
 import styled from 'styled-components';
-import { button } from 'styles/Shared';
 
-export const FormWrap = styled.form`
-  padding: 0 20px;
-`;
-
-export const TextBox = styled.textarea`
-  resize: none;
-  width: 100%;
-  height: 268px;
-  padding: 14px 16px;
-  border: 1px solid ${({ theme }) => theme.color.gray_2};
-  border-radius: 6px;
+export const PhotoInput = styled.input`
+  display: none;
 `;
 
 export const FieldWrap = styled.fieldset`
@@ -20,10 +10,6 @@ export const FieldWrap = styled.fieldset`
   &.hasPhoto {
     padding-bottom: 21px;
   }
-`;
-
-export const PhotoInput = styled.input`
-  display: none;
 `;
 
 export const AddPhotoBtn = styled.button`
@@ -55,9 +41,4 @@ export const ImgWrap = styled.div`
     height: 71px;
     border-radius: 4px;
   }
-`;
-
-export const SubmitButton = styled.button`
-  ${button}
-  margin-bottom: 34px;
 `;
