+++ src/components/FeedAdd/SelectedCategory/SelectedCategory.jsx
@@ -3,16 +3,14 @@ import { Icon } from 'components/Common';
 import { ReactComponent as Question } from 'assets/icons/question_mark.svg';
 import * as S from './styles';
 
-export const SelectedCategory = ({
-  selecting,
-  setToggleSelect,
-  currentCategory
-}) => {
+export const SelectedCategory = ({ setToggleSelect, currentCategory }) => {
   return (
     <S.CategoryBox>
       <S.CategoryImg
-        className={`${selecting && 'onSelecting'}
-          ${currentCategory !== '카테고리 선택' && 'onSelected'}`}
+        className={`${
+          (currentCategory === '카테고리 선택' && 'onSelecting') ||
+          (currentCategory !== '카테고리 선택' && 'onSelected')
+        }`}
       >
         {currentCategory === '카테고리 선택' ? (
           <Icon Icon={Question} size={18} color={'#fff'} />
@@ -21,7 +19,7 @@ export const SelectedCategory = ({
 
       <S.CategoryNameBtn
         onClick={setToggleSelect}
-        className={`${selecting && 'onSelecting'}`}
+        className={`${currentCategory === '카테고리 선택' && 'onSelecting'}`}
       >
         {currentCategory}
       </S.CategoryNameBtn>
