+++ src/components/FeedAdd/index.js
@@ -1,3 +1,3 @@
 export { SelectedCategory } from './SelectedCategory/SelectedCategory';
 
-export { ContentsBox } from './ContentsBox/ContentsBox';
+export { FeedForm } from './FeedForm/FeedForm';
